requirements.txt
boto3==1.12.26
botocore==1.15.26
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chart-studio==1.0.0
click==7.1.1
cycler==0.10.0
docutils==0.15.2
Flask==1.1.1
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googlemaps==4.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mpld3==0.3
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
plotly==4.5.4
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
smart-open==1.10.0
stop-words==2018.7.23
twitter==1.18.0
urllib3==1.25.8
Werkzeug==1.0.0
wordcloud==1.6.0
